+++ zarrs/tests/data/v2_generate.py
@@ -1,3 +1,14 @@
+#!/usr/bin/env -S uv run
+# /// script
+# requires-python = ">=3.12"
+# dependencies = [
+#     "zarr==3.0.1",
+#     "numcodecs==0.15.0",
+#     "zfpy==1.0.1",
+#     "pcodec==0.3.2",
+# ]
+# ///
+
 import zarr
 import numpy as np
 from numcodecs import Blosc, GZip, BZ2, ZFPY, PCodec, Zstd